compiles with wicket 6.0.0

+++ de.wicketpraxis--pom/webapp/src/main/java/de/wicketpraxis/web/thema/komponenten/basis/repeater/DataGridViewPage.java
@@ -57,7 +57,7 @@ public class DataGridViewPage extends WebPage {
 
 			@Override
 			public void onClick() {
-				int page = datagrid.getCurrentPage() + 1;
+				long page = datagrid.getCurrentPage() + 1;
 				if (page < datagrid.getPageCount())
 					datagrid.setCurrentPage(page);
 			}
